expense planning/Internal-hackathon/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Login==0.6.3
python-dotenv==1.0.1
Werkzeug==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
reportlab==4.2.2
